#173 move file operations out of src/agenda
