+++ src/money/test_treasury/test_treasury_voice_rank.py
@@ -1,7 +1,7 @@
 from src._instrument.file import save_file, delete_dir
 from src.agenda.party import partyunit_shop
 from src.agenda.agenda import agendaunit_shop
-from src.agenda.meld_files import get_file_names_in_voice_rank_order
+from src.change.meld_files import get_file_names_in_voice_rank_order
 from src.change.listen import create_job_file_from_role_file
 from src.money.money import moneyunit_shop
 from src.money.examples.econ_env import env_dir_setup_cleanup, get_texas_filehub
